refactor(cuda-graphs): move partial graphs into component
